create reviews sort bar component
